third_party/web_platform_tests/workers/interfaces/WorkerUtils/WindowTimers/002.html

<!-- Pulled from chromium at commit: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/external/wpt/workers/interfaces/WorkerUtils/WindowTimers/002.html;drc=e5b8fefe2aae5674dc00817d444a50f67bdb6ea0 -->
<!doctype html>
<title>clearTimeout</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
  async_test(function () {
    var worker = new Worker('002.js');
    var gotMessage = false;
    worker.onmessage = function () { gotMessage = true; };
    setTimeout(this.step_func(function () { assert_false(gotMessage); this.done(); }), 100);
  });
</script>
